requirements.txt
python-dotenv>=0.12.0,<1.0.0
Flask>=1.1.1,<2.0.0
Flask-SQLAlchemy>=2.4.1,<3.0.0
Flask-Session>=0.3.1,<1.0.0
Flask-Login>=0.5.0,<1.0.0
Flask-Migrate>=2.5.3,<3.0.0
gunicorn>=20.0.4,<21.0.0
requests-oauthlib>=1.3.0,<2.0.0
requests>=2.23.0,<3.0.0
marshmallow==2.16.3
psycopg2>=2.8.4,<3.0.0
flask-jwt-extended>=3.24.1,<4.0.0
